requirements.txt
# Aplicación Streamlit
streamlit>=1.28.0

# Procesamiento de datos
pandas>=2.0.0
numpy>=1.24.0

# Procesamiento de texto y NLP
nltk>=3.8.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.0
python-Levenshtein>=0.21.0

# Análisis bibliométrico
bibtexparser>=2.0.0

# Visualización
matplotlib>=3.7.0
seaborn>=0.12.0
wordcloud>=1.9.0
plotly>=5.15.0

# Geografía (Requerimiento 5)
geopandas>=0.13.0
pycountry>=23.0.0

# Clustering (Requerimiento 4)
scipy>=1.11.0

# PDF (Requerimiento 2 y 5)
reportlab>=4.0.0
Pillow>=10.0.0

# APIs y descarga (Requerimiento 1)
requests>=2.31.0

# Utilidades
pathlib2>=2.3.7; python_version < "3.4"

